refactor SchemaBuilder
